Update with_form to with_pane in TabListComponent
