+++ app/components/elements/tab_form/pane_component.html.erb
@@ -12,7 +12,7 @@
                 <% if render_footer? %>
                     <div>
                         <%= render Elements::Forms::SubmitComponent.new(form:, label: 'Save as draft', variant: :'outline-primary') %>
-                        <%= render Elements::NextButtonComponent.new(tab_id:) %>
+                        <%= render Elements::TabForm::NextButtonComponent.new(tab_id:) %>
                     </div>
                 <% end %>
             </div>
