--- app/components/elements/tab_list_component.html.erb
@@ -1,8 +1,0 @@
-<%= tag.div class: classes, 'data-controller': 'tab-error' do %>
-  <%= tag.div class: tab_classes, role: 'tablist', 'aria-orientation': 'vertical' do |component| %>
-    <% tabs.each do |tab| %>
-      <%= tab %>
-    <% end %>
-  <% end %>
-  <%= form %>
-<% end %>
